refactor: extract observation store schema seam
